+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 3.5.11
 * Ender Dragon now will now land on the exact center of the podium, to prevent begin too low and hitting players *in* the podium
 * Wither Charge attack damage is now configurable
 * Fixed larvae dropping ancient debris and nether star shard